backend/requirements.txt
Flask>=2.3.2
Flask-CORS>=4.0.0
torch>=2.0.1,<2.1.0
torchvision>=0.15.2,<0.16.0
pillow>=9.5.0,<10.0.0
numpy>=1.24.3,<1.25
typing-extensions>=4.5.0,<5.0
Werkzeug>=2.2.3,<3.0
protobuf<3.20,>=3.9.2
python-dotenv>=1.0.0
gunicorn>=21.2.0
requests>=2.31.0
opencv-python>=4.8.0
scikit-image>=0.21.0
